+++ backend/app/schemas/guide_content.py
