+++ backend/app/services/gpt.py
@@ -38,7 +38,7 @@ def _compose_prompt(
 
     return (
         "あなたは観光ガイドです。"
-        f"{tone}、1分（200字）程度のスピーチ台本を作成してください。\n"
+        f"{tone}、1〜2分（200〜400字）程度のスピーチ台本を作成してください。\n"
         "構成は【概要→見どころ→歴史/豆知識→おすすめの楽しみ方→アクセス/注意点】。\n"
         "事実ベースで、固有名詞や数字はできるだけ具体的に。誇張は避けてください。\n"
         f"- 場所名: {name}\n"
